Move chrono literals to google/cloud/testing_util.
These literals are generally useful when writing tests, move
them to the shared directory.
