+++ google/cloud/bigtable/bigtable_client_testing.bzl
@@ -1,6 +1,5 @@
 # DO NOT EDIT -- GENERATED BY CMake -- Change the CMakeLists.txt file if needed
 bigtable_client_testing_HDRS = [
-    "testing/chrono_literals.h",
     "testing/embedded_server_test_fixture.h",
     "testing/internal_table_test_fixture.h",
     "testing/mock_admin_client.h",
