+++ google/cloud/bigtable/mutations_test.cc
@@ -13,12 +13,12 @@
 // limitations under the License.
 
 #include "google/cloud/bigtable/mutations.h"
-#include "google/cloud/bigtable/testing/chrono_literals.h"
+#include "google/cloud/testing_util/chrono_literals.h"
 #include <google/rpc/error_details.pb.h>
 #include <gmock/gmock.h>
 
 namespace bigtable = google::cloud::bigtable;
-using namespace bigtable::chrono_literals;
+using namespace google::cloud::testing_util::chrono_literals;
 
 /// @test Verify that SetCell() works as expected.
 TEST(MutationsTest, SetCell) {
