+++ google/cloud/bigtable/polling_policy_test.cc
@@ -13,8 +13,8 @@
 // limitations under the License.
 
 #include "google/cloud/bigtable/polling_policy.h"
-#include "google/cloud/bigtable/testing/chrono_literals.h"
 #include "google/cloud/testing_util/check_predicate_becomes_false.h"
+#include "google/cloud/testing_util/chrono_literals.h"
 #include <gtest/gtest.h>
 #include <chrono>
 #include <thread>
@@ -27,7 +27,7 @@ grpc::Status CreatePermanentError() {
 }
 
 namespace bigtable = google::cloud::bigtable;
-using namespace bigtable::chrono_literals;
+using namespace google::cloud::testing_util::chrono_literals;
 auto const kLimitedTimeTestPeriod = 50_ms;
 auto const kLimitedTimeTolerance = 10_ms;
