+++ google/cloud/bigtable/table_admin_test.cc
@@ -15,8 +15,8 @@
 #include "google/cloud/bigtable/table_admin.h"
 #include "google/cloud/bigtable/grpc_error.h"
 #include "google/cloud/bigtable/internal/make_unique.h"
-#include "google/cloud/bigtable/testing/chrono_literals.h"
 #include "google/cloud/bigtable/testing/mock_admin_client.h"
+#include "google/cloud/testing_util/chrono_literals.h"
 #include <google/protobuf/text_format.h>
 #include <google/protobuf/util/message_differencer.h>
 #include <gmock/gmock.h>
@@ -258,7 +258,7 @@ TEST_F(TableAdminTest, ListTablesUnrecoverableFailures) {
  */
 TEST_F(TableAdminTest, ListTablesTooManyFailures) {
   using namespace ::testing;
-  using namespace bigtable::chrono_literals;
+  using namespace google::cloud::testing_util::chrono_literals;
 
   bigtable::TableAdmin tested(
       client_, "the-instance", bigtable::LimitedErrorCountRetryPolicy(3),
@@ -283,7 +283,7 @@ TEST_F(TableAdminTest, ListTablesTooManyFailures) {
 /// @test Verify that `bigtable::TableAdmin::Create` works in the easy case.
 TEST_F(TableAdminTest, CreateTableSimple) {
   using namespace ::testing;
-  using namespace bigtable::chrono_literals;
+  using namespace google::cloud::testing_util::chrono_literals;
 
   bigtable::TableAdmin tested(client_, "the-instance");
 
@@ -372,7 +372,7 @@ TEST_F(TableAdminTest, CopyConstructibleAssignableTest) {
  */
 TEST_F(TableAdminTest, CopyConstructibleAssignablePolicyTest) {
   using namespace ::testing;
-  using namespace bigtable::chrono_literals;
+  using namespace google::cloud::testing_util::chrono_literals;
 
   bigtable::TableAdmin tested(
       client_, "the-construct-instance",
@@ -404,7 +404,7 @@ TEST_F(TableAdminTest, CopyConstructibleAssignablePolicyTest) {
 /// @test Verify that `bigtable::TableAdmin::GetTable` works in the easy case.
 TEST_F(TableAdminTest, GetTableSimple) {
   using namespace ::testing;
-  using namespace bigtable::chrono_literals;
+  using namespace google::cloud::testing_util::chrono_literals;
 
   bigtable::TableAdmin tested(client_, "the-instance");
   std::string expected_text = R"""(
@@ -428,7 +428,7 @@ TEST_F(TableAdminTest, GetTableSimple) {
  */
 TEST_F(TableAdminTest, GetTableUnrecoverableFailures) {
   using namespace ::testing;
-  using namespace bigtable::chrono_literals;
+  using namespace google::cloud::testing_util::chrono_literals;
 
   bigtable::TableAdmin tested(client_, "the-instance");
   EXPECT_CALL(*client_, GetTable(_, _, _))
@@ -450,7 +450,7 @@ TEST_F(TableAdminTest, GetTableUnrecoverableFailures) {
  */
 TEST_F(TableAdminTest, GetTableTooManyFailures) {
   using namespace ::testing;
-  using namespace bigtable::chrono_literals;
+  using namespace google::cloud::testing_util::chrono_literals;
 
   bigtable::TableAdmin tested(
       client_, "the-instance", bigtable::LimitedErrorCountRetryPolicy(3),
@@ -512,7 +512,7 @@ TEST_F(TableAdminTest, DeleteTableFailure) {
  */
 TEST_F(TableAdminTest, ModifyColumnFamilies) {
   using namespace ::testing;
-  using namespace bigtable::chrono_literals;
+  using namespace google::cloud::testing_util::chrono_literals;
   using google::protobuf::Empty;
 
   bigtable::TableAdmin tested(client_, "the-instance");
@@ -545,7 +545,7 @@ modifications {
  */
 TEST_F(TableAdminTest, ModifyColumnFamiliesFailure) {
   using namespace ::testing;
-  using namespace bigtable::chrono_literals;
+  using namespace google::cloud::testing_util::chrono_literals;
 
   bigtable::TableAdmin tested(client_, "the-instance");
   EXPECT_CALL(*client_, ModifyColumnFamilies(_, _, _))
@@ -744,7 +744,7 @@ TEST_F(TableAdminTest, CheckConsistencyFailure) {
  */
 TEST_F(TableAdminTest, AsyncCheckConsistencySimple) {
   using namespace ::testing;
-  using namespace bigtable::chrono_literals;
+  using namespace google::cloud::testing_util::chrono_literals;
 
   bigtable::TableAdmin tested(client_, "the-async-instance");
   std::string expected_text = R"""(
@@ -780,7 +780,7 @@ TEST_F(TableAdminTest, AsyncCheckConsistencySimple) {
  */
 TEST_F(TableAdminTest, AsyncCheckConsistencyFailure) {
   using namespace ::testing;
-  using namespace bigtable::chrono_literals;
+  using namespace google::cloud::testing_util::chrono_literals;
 
   bigtable::TableAdmin tested(client_, "the-async-instance");
   EXPECT_CALL(*client_, CheckConsistency(_, _, _))
@@ -807,7 +807,7 @@ TEST_F(TableAdminTest, AsyncCheckConsistencyFailure) {
  */
 TEST_F(TableAdminTest, GetSnapshotSimple) {
   using namespace ::testing;
-  using namespace bigtable::chrono_literals;
+  using namespace google::cloud::testing_util::chrono_literals;
 
   bigtable::TableAdmin tested(client_, "the-instance");
   std::string expected_text = R"""(
@@ -831,7 +831,7 @@ TEST_F(TableAdminTest, GetSnapshotSimple) {
  */
 TEST_F(TableAdminTest, GetSnapshotUnrecoverableFailures) {
   using namespace ::testing;
-  using namespace bigtable::chrono_literals;
+  using namespace google::cloud::testing_util::chrono_literals;
 
   bigtable::TableAdmin tested(client_, "the-instance");
   EXPECT_CALL(*client_, GetSnapshot(_, _, _))
@@ -855,7 +855,7 @@ TEST_F(TableAdminTest, GetSnapshotUnrecoverableFailures) {
  */
 TEST_F(TableAdminTest, GetSnapshotTooManyFailures) {
   using namespace ::testing;
-  using namespace bigtable::chrono_literals;
+  using namespace google::cloud::testing_util::chrono_literals;
 
   bigtable::TableAdmin tested(
       client_, "the-instance", bigtable::LimitedErrorCountRetryPolicy(3),
@@ -968,7 +968,7 @@ TEST_F(TableAdminTest, ListSnapshots_SimpleList) {
  */
 TEST_F(TableAdminTest, ListSnapshots_RecoverableFailure) {
   using namespace ::testing;
-  using namespace bigtable::chrono_literals;
+  using namespace google::cloud::testing_util::chrono_literals;
 
   bigtable::TableAdmin tested(client_, "the-instance");
   auto mock_recoverable_failure =
