+++ google/cloud/bigtable/tests/data_integration_test.cc
@@ -13,8 +13,8 @@
 // limitations under the License.
 
 #include "google/cloud/bigtable/internal/endian.h"
-#include "google/cloud/bigtable/testing/chrono_literals.h"
 #include "google/cloud/bigtable/testing/table_integration_test.h"
+#include "google/cloud/testing_util/chrono_literals.h"
 #include "google/cloud/testing_util/init_google_mock.h"
 
 namespace {
@@ -104,7 +104,7 @@ void DataIntegrationTest::BulkApply(bigtable::Table& table,
 }
 }  // anonymous namespace
 
-using namespace bigtable::chrono_literals;
+using namespace google::cloud::testing_util::chrono_literals;
 
 TEST_F(DataIntegrationTest, TableApply) {
   std::string const table_id = RandomTableId();
