+++ google/cloud/google_cloud_cpp_testing.bzl
@@ -1,6 +1,7 @@
 # DO NOT EDIT -- GENERATED BY CMake -- Change the CMakeLists.txt file if needed
 google_cloud_cpp_testing_HDRS = [
     "testing_util/check_predicate_becomes_false.h",
+    "testing_util/chrono_literals.h",
     "testing_util/environment_variable_restore.h",
     "testing_util/init_google_mock.h",
 ]
